Separate browse into logical components
