+++ application/src/View/Helper/Browse.php
@@ -1,28 +1,26 @@
 <?php
 namespace Omeka\View\Helper;
 
-use Omeka\Api\Manager as ApiManager;
-use Omeka\Api\Representation\AbstractEntityRepresentation;
-use Omeka\ColumnType\ColumnTypeInterface;
 use Laminas\Form\Element;
 use Laminas\ServiceManager\ServiceLocatorInterface;
 use Laminas\View\Helper\AbstractHelper;
+use Omeka\Api\Manager as ApiManager;
+use Omeka\Api\Representation\AbstractEntityRepresentation;
+use Omeka\ColumnType\ColumnTypeInterface;
+use Omeka\Stdlib\Browse as BrowseService;
 
 class Browse extends AbstractHelper
 {
     protected ServiceLocatorInterface $services;
 
-    protected array $columnDefaults;
-    protected array $browseDefaults;
-    protected array $sortDefaults;
-
     public function __construct(ServiceLocatorInterface $services)
     {
         $this->services = $services;
-        $config = $services->get('Config');
-        $this->columnDefaults = $config['column_defaults'];
-        $this->browseDefaults = $config['browse_defaults'];
-        $this->sortDefaults = $config['sort_defaults'];
+    }
+
+    public function getBrowseService() : BrowseService
+    {
+        return $this->services->get('Omeka\Browse');
     }
 
     /**
@@ -31,7 +29,6 @@ class Browse extends AbstractHelper
      * Pass a resource type (string) to use configured/default sorts. Otherwise,
      * pass a sort configuration (array).
      *
-     * @see self::getSortConfig()
      * @param string|array $resourceTypeOrSortConfig
      */
     public function renderSortSelector($resourceTypeOrSortConfig) : string
@@ -39,7 +36,7 @@ class Browse extends AbstractHelper
         $view = $this->getView();
         $context = $view->status()->isAdminRequest() ? 'admin' : 'public';
         if (is_string($resourceTypeOrSortConfig)) {
-            $sortConfig = $this->getSortConfig($context, $resourceTypeOrSortConfig);
+            $sortConfig = $this->getBrowseService()->getSortConfig($context, $resourceTypeOrSortConfig);
         } elseif (is_array($resourceTypeOrSortConfig)) {
             $sortConfig = $resourceTypeOrSortConfig;
         } else {
@@ -57,79 +54,6 @@ class Browse extends AbstractHelper
     }
 
     /**
-     * Get the sort configuration.
-     *
-     * The sort configuration is an array:
-     *
-     * [
-     *   'sort_by_query_param' => 'Sort by label',
-     *   'another_sort_by_query_param' => 'Another sort by label',
-     * ]
-     */
-    public function getSortConfig(string $context, string $resourceType) : array
-    {
-        $sortConfig = [];
-        // Include sorts from user-configured columns.
-        foreach ($this->getColumnsData($context, $resourceType) as $columnData) {
-            if (!$this->columnTypeIsKnown($columnData['type'])) {
-                continue; // Skip unknown column types.
-            }
-            $columnType = $this->getColumnType($columnData['type']);
-            $sortBy = $columnType->getSortBy($columnData);
-            if (!$sortBy) {
-                continue; // This column cannot be sorted.
-            }
-            $sortConfig[$sortBy] = $this->getHeader($columnData);
-        }
-        // Include default sorts that are not configured.
-        $sortDefaults = $this->sortDefaults[$context][$resourceType] ?? [];
-        foreach ($sortDefaults as $sortBy => $label) {
-            if (!isset($sortConfig[$sortBy])) {
-                $sortConfig[$sortBy] = $label;
-            }
-        }
-        // Include any other sorts added by the sort-config event.
-        $eventManager = $this->services->get('EventManager');
-        $args = $eventManager->prepareArgs([
-            'context' => $context,
-            'resourceType' => $resourceType,
-            'sortConfig' => $sortConfig,
-        ]);
-        $eventManager->trigger('sort-config', null, $args);
-        $sortConfig = $args['sortConfig'];
-        natsort($sortConfig);
-        return $sortConfig;
-    }
-
-    /**
-     * Get the browse configuration.
-     *
-     * The browse configuration is an array with three elements:
-     *   1. The default sort_by value
-     *   2. The default sort_order value
-     *   3. The default page value
-     */
-    public function getBrowseConfig(string $context, string $resourceType, ?int $userId = null) : array
-    {
-        $view = $this->getView();
-        $userSettings = $this->services->get('Omeka\Settings\User');
-        // First, get the user-configured browse defaults, if any. Set the
-        // defaults from the config file if they're not configured or malformed.
-        $browseDefaultsSetting = sprintf('browse_defaults_%s_%s', $context, $resourceType);
-        $browseConfig = $userSettings->get($browseDefaultsSetting, null, $userId);
-        if (!is_array($browseConfig) || !isset($browseConfig[0]) || !is_string($browseConfig[0])) {
-            $browseConfig = $this->browseDefaults[$context][$resourceType] ?? [null, 'desc', 1];
-        }
-        // Standardize the defaults before returning.
-        $browseConfig = [
-            $browseConfig[0] ?? null,
-            $browseConfig[1] ?? 'desc',
-            $browseConfig[2] ?? 1,
-        ];
-        return $browseConfig;
-    }
-
-    /**
      * Get the header row for a resource type.
      */
     public function renderHeaderRow(string $resourceType) : string
@@ -137,8 +61,8 @@ class Browse extends AbstractHelper
         $view = $this->getView();
         $context = $view->status()->isAdminRequest() ? 'admin' : 'public';
         $headerRow = [];
-        foreach ($this->getColumnsData($context, $resourceType) as $columnData) {
-            if (!$this->columnTypeIsKnown($columnData['type'])) {
+        foreach ($this->getBrowseService()->getColumnsData($context, $resourceType) as $columnData) {
+            if (!$this->getBrowseService()->columnTypeIsKnown($columnData['type'])) {
                 continue; // Skip unknown column types.
             }
             $headerRow[] = sprintf(
@@ -159,7 +83,7 @@ class Browse extends AbstractHelper
     public function getHeader(array $columnData) : string
     {
         $view = $this->getView();
-        $columnType = $this->getColumnType($columnData['type']);
+        $columnType = $this->getBrowseService()->getColumnType($columnData['type']);
         if (isset($columnData['header']) && '' !== trim($columnData['header'])) {
             return $columnData['header'];
         }
@@ -174,8 +98,8 @@ class Browse extends AbstractHelper
         $view = $this->getView();
         $context = $view->status()->isAdminRequest() ? 'admin' : 'public';
         $contentRow = [];
-        foreach ($this->getColumnsData($context, $resourceType) as $columnData) {
-            if (!$this->columnTypeIsKnown($columnData['type'])) {
+        foreach ($this->getBrowseService()->getColumnsData($context, $resourceType) as $columnData) {
+            if (!$this->getBrowseService()->columnTypeIsKnown($columnData['type'])) {
                 continue; // Skip unknown column types.
             }
             $contentRow[] = sprintf(
@@ -196,41 +120,8 @@ class Browse extends AbstractHelper
     public function getContent(AbstractEntityRepresentation $resource, array $columnData) : ?string
     {
         $view = $this->getView();
-        $columnType = $this->getColumnType($columnData['type']);
+        $columnType = $this->getBrowseService()->getColumnType($columnData['type']);
         return  $columnType->renderContent($view, $resource, $columnData) ?? $columnData['default'];
-    }
-
-    /**
-     * Get data for all columns.
-     */
-    public function getColumnsData(string $context, string $resourceType, ?int $userId = null) : array
-    {
-        $view = $this->getView();
-        $userSettings = $this->services->get('Omeka\Settings\User');
-        // First, get the user-configured columns data, if any. Set the default
-        // if data is not configured or malformed.
-        $userColumnsSetting = sprintf('columns_%s_%s', $context, $resourceType);
-        $userColumnsData = $userSettings->get($userColumnsSetting, null, $userId);
-        if (!is_array($userColumnsData) || !$userColumnsData) {
-            $userColumnsData = $this->columnDefaults[$context][$resourceType] ?? [];
-        }
-        // Standardize the data before returning.
-        $columnsData = [];
-        foreach ($userColumnsData as $index => $userColumnData) {
-            if (!is_array($userColumnData)) {
-                // Skip columns that are not an array.
-                continue;
-            }
-            if (!isset($userColumnData['type'])) {
-                // Skip columns without a type.
-                continue;
-            }
-            // Add required keys if not present.
-            $userColumnData['default'] ??= null;
-            $userColumnData['header'] ??= null;
-            $columnsData[] = $userColumnData;
-        }
-        return $columnsData;
     }
 
     /**
@@ -283,7 +174,7 @@ class Browse extends AbstractHelper
             'value' => $columnType->getLabel(),
         ]);
         $columnForm[] = $view->formRow($columnTypeInput);
-        if ($this->columnTypeIsKnown($columnData['type'])) {
+        if ($this->getBrowseService()->columnTypeIsKnown($columnData['type'])) {
             $headerInput = $formElements->get(Element\Text::class);
             $headerInput->setName('column_header');
             $headerInput->setOptions([
@@ -308,22 +199,5 @@ class Browse extends AbstractHelper
         }
         $columnForm[] = $columnType->renderDataForm($view, $columnData);
         return implode($columnForm);
-    }
-
-    /**
-     * Get a column type by name.
-     */
-    public function getColumnType(string $columnType) : ColumnTypeInterface
-    {
-        return $this->services->get('Omeka\ColumnTypeManager')->get($columnType);
-    }
-
-    /**
-     * Is this column type known?
-     */
-    public function columnTypeIsKnown(string $columnType) : bool
-    {
-        $columnTypes = $this->services->get('Omeka\ColumnTypeManager');
-        return $columnTypes->has($columnType);
     }
 }
